Adding initial site
This is a work in progress commit. It includes all base provisioning and Pristine template related files, as well as the under-construction Jekyll site.

The main site collections have been added and work on a dynamic index has began. A couple of pages have been added to the 'core styles' collection to test this, and one of these have been developed with 'sections' (which will either be directly in-line, or use includes, I haven't decided yet).

Main navigation is in-place but will be improved to be more dynamic.