starting to work on the main home page
This is all fairly custom so has been converted back to HTML for ease of structuring etc.

The proper BAS logo is now used, but needs to have the surrounding ring added.
Other styling tweaks are also needed, but this represents all the HTML elements needed.

A future version of the style kit will include BAS logos in various forms, which this site can then use. Therefore the included BAS logo is considered temporary.