Improving layout/styling of homepage
Includes a tempary set of styles for alerts (which will be formalised by an updated alerts component as part of the next Style Kit release).

Also includes a better copyright notice for images (fixing all of the issues with the previous implementation where the notice placement was inconsistent) - again this will become standard as part of the next Style Kit release.

Various other tweaks are included, such as better phrasing for the BETA banner.