Adding better styling for sidebar navigation in topics
This needs some more thinking in terms of the general implementation - for example, removing the sidebar completely and using the primary navigation to move between collections/topics and a static TOC for navigating to sections.

Since the sidebar is now static and doesn't track page position there is no real difference between an inline TOC and a sidebar - except the sidebar takes up more page space, and looks odd when no longer visible.

The TOC approach works well for Wikipedia - and is being implemented in the new Bootstrap docs so wouldn't be an issue for expected UX.