Adding a staging environment
This uses developments from the BAS-AWS project for defining IAM users and policies through Terraform. An S3 bucket is currently used, but this will migrate to a combined 'pages' project when that's ready.

This also adds CI/CD services for building and deploying updates - note the production site won't be covered by this project until the 0.2.0 release of the Style Kit is ready.