Fixing URL structure
File ordering shouldn't be represented in output URLs. This means either slugs need to be added to each page, or, the ordering numbers need to be removed.

There is less of a need to structure topics within a collection to fit some sort of narative going down the page - as each page is only concerned with a specific topic.

The permalink structure has also been corrected to ensure we have pretty URLs for collection items.