Adding support for fetching remote Jekyll data files
This uses a plugin from 18F which requires a number of additional Gems to be installed.

These Gems are specified in a Gemfile and installed with Bundler, which is currently installed manually but will be included in future versions of the Ruby2 BARC role.

Currently this plugin is used to fetch the SRI values for the Style Kit from the BAS CDN. This is slight problem as this is the same location as the distribution files themselves. Since this feature is designed to protect against a CDN compromise this represents a problem - if you can hack the assets, you could also hack the SRI values.

There are workarounds to this problem - we can host the SRI file somewhere else (say the BAS website) - however since this file is only picked up at build-time - and not on every browser request, this shouldn't be a significant risk - as if you could modify the CDN you could probably modify the site source code anyway.

This is a limitation though and should be addressed in the medium term.