Adding Jira issue plugin
This provides a new tag for referencing a Jira issue by its issue id. Details of the issue, specifically its status, are retrieved from the Jira API and displayed next to a link to the given issue.

This is similar to the functionality Confluence provides, and will be useful in primarily internal facing documentation for other projects.

This plugin is still considered proof of concept, but is designed to the open-sourced, with all configuration set within the Jekyll config file.

There are still some rough edges, setting the account password as a variable for example is annoying in development environments. It may be better to disable the plugin in these cases as the information is only really useful when actually published.

The markup used by the plugin will be added as a new component to the style kit for future standardisation.