Adding environment support for Jira plugin
This means in development the API call is skipped and a faked response returned.

This ensures local builds remain quick and removes the need for keeping the Jira account password lying around.

The Jekyll environment is set to production for CI/CD so this will be tested properly on each commit of any branch for extra protection (i.e. in case the plugin stops working fetching issue details from Jira).