Adding core layout documentation
Initially this layout topic would have included sections on the grid, but this would have lead to a very long, and dense, page.

I therefore think its better to continue to have a layout and grid page seperately. The layout page will focus on the underlying elements, namely containers and media queries, whereas grids is more practical, as an implementation of these queries.

The grid documentation is yet to be added, but is not as difficult since it is more practical than theoretical.