Fixing grid class examples
In most casesthe grid will be used for layout in examples, rather than being the focus of the example itself.

There are cases where this is needed however (such as the grid examples).

Previously any use of a grid class would add an outline and fill, which looks odd when used for layout.

These extra styles are now scoped so they can still be used for grid layouts, but remain 'invisible' when used in all other examples.