Adding testing styles
These will need to be formalised into a proper conventional colour and icon, and variants of components such as alerts and navbars.

The purpose of these styles is to visually distingish between testing and production versions of websites and web-apps.

In additional to site wide elements, context specific warnings/notices would be needed on confirmation screens for example (e.g. ‘the DOI you are about to create will not really exist as this is a testing website’ etc.).

See: https://trello.com/c/7k80ugBx/25-add-formal-testing-style for more.