Adding better docker support
More work is needed to create a private docker image which contains Gems, but not the site contents.

To solve the latter we can use a volume to expose the local file system as '/usr/src/app/site'.

To solve the former we can use GitLab.

For CI the Jekyll-Deployment-Image should be used instead (which is already public).