Updating CI to use main project image
By using ‘before_script’ to change into the site directory we should be able to use the main Docker image for this project in CI.

Previously GitLab needed its own image as needed to run ‘cd site’ first, but couldn’t as the jekyll-image this project’s image depends on sets ‘jekyll’ as the entry command - which GitLab cannot currently override.

This may still not work however as we are still setting a volume and working directory within the Docker image.