Cleaning up GitLab CI implementation
GitLab CI is now working using the project image. With a compromise that the entrypoint defined by the ‘antarctica/jekyll-image’ image has to be overriden, and the ‘jekyll’ command defined within the CMD.

This goes against Docker’s best practices, but can still be overriden, and so isn’t the worst hack in the world.