Adding ‘pack’ stage/job to CD
This will eventually take the build produced by CI and archive into a zip. For now we are testing where artifacts (or dependencies) are available from.

Wen this can produce Zips, this new stage can be used by later stages to upload to the BAS Packages Service for example.

The longer term aim of this work is to replicate much of what the jekyll-deployment-image does, but in a more generic way. For example, if the site files are already built, then the deployment aspects don’t need Ruby anymore.

This means for the deployment stages we are free to use more minimal deployment image (rather than build and deployment), using just Python/AWS-CLI.

The ability to use different Docker images at different stages within a CI/CD workflow is very useful for keeping images generic, with project specific tasks used to glue these together.

This is in contrast to BitBucket where we can only define one image, and so have an ‘all or nothing’ approach.