Updating provisioning
Adds a new ‘site-all’ environment for defining an IAM user used by both ‘site-stage’ and ‘site-prod’.

GitLab will use this user to upload snapshots and deploy changes to websites.

Documentation for creating the staging environment has been rewritten and placed after local development since this is only really needed as a reference once its been performed.

The existing ‘site-stage’ environment has been updated to remove references to Atlas and to use data sources introduced in recent Terraform versions.

The existing IAM user defined for the ‘site-stage’ environment is removed, and will break the existing Semaphore based CD process. However as this has been replaced with GitLab this is fine and intentional.

Inline IAM polices are still defined in this environment, and attached to the ‘common’ project user defined in the ‘site-all’ environment.