Tidying up topic sections
This adds a new snippet (really a renamed snippet) which combines the ‘back-to-top’ and ‘section-metadata’ snippets together, such that there’s much greater consistency.

This new ‘topic-section-metadata’ snippet is to be used everywhere. If version information is needed for a section, it can be safety omitted.

This new snippet removes the ‘bootstrap/custom’ property - since this is going to become meaningless if we add the global namespace (i.e. ‘bak-foo’).

This new snippet includes responsibility for generating titles and anchor links, and means the title is specified once as an input to this snippet. This in theory means the two cannot get out of sync. This change does remove syntax highlighting for headers, which might become a problem in future and need to be reviewed.

A permalink is now automatically generated for each section to allow specific sections to be referenced. This could be done before using the link in topic collection indexes, or topic page contents lists, but this is a far more discoverable option.

Because this snippet is used pretty much everywhere, this commit includes changes to almost every documentation page.