Removing Gem lock file
Ensures the latest versions of Gems are installed on each image build.

Since the same image is used in development and CI/CD there is no danger from then on of Gem versions changing (which is what the Gem lock file is designed to protect against).

Doing it this way just means we have to remove the Gem lock file, update, and extract this back out to the repository (as its added to the container explicitly, not mounted as a volume).