Using Jekyll environment directly
Previously we used the branch name (set by CI environments as an environment variable) as a proxy over the JEKYLL_ENV variable, which serves the same purpose, and is conventional for Jekyll sites.

Since we can trivially set this variable in Docker compose (for development), and GitLab CI (for staging/production) there is no need to make things more complicated with another variable doing essentially the same thing (but non-standidly).

Even though the default for this variable is ‘development’ we set it explicitly to make it obvious.