Fixing outline error
This was an annoying bug caused by the ‘tabindex=“-1”’ fix applied to improve accessibility. A side effect of this was for non-interactive elements to have an outline on focus as if they were interactive.

This CSS fix removes this outline and will be incorporated into the main Style Kit in due course.

This accessibility fix was not universally applied to topic indexes which is also fixed.