Adding logos core component
Details the use of the BAS Logo and OGL Symbol.

The format of this might change as there are lots of alerts which are currently oversized.

Components that use logos have been updated to reference this. The distribution getting started page has been updated with a logos section similar to web-fonts.