Adding notice about outlines disabled for body elements
This is caused by trying to do the right thing (using skip navigation links), which need a workaround to implement across browsers (tabindex=-1).

This workaround has side-effects however by making it look like the element the skip navigation link points to is interactive (as tab-index is designed for interactive elements such as inputs). This is not ideal as it looks bad and breaks the users expectations.

To workaround this bug, we hide the outline which doesn’t get rid of the problem - but does stop people noticing it.