Updating header layout components
A new abstraction ‘site-header’ is added (which mirrors ‘site-footer’) and contains the cookie notice, main navigation and development phase banner.

This also resolves a long standing layout problem we had with these components not aligning correctly (and still isn’t completely solved).

Note: For now this also removes the testing banner shown on the testing site (the navbar icon is still shown).