Overriding the Style Kit Theme CSS include
In this case, and possibly only this case, we will always want to load a custom version of the Style Kit, i.e. separate to the version included in the theme.

Currently the theme doesn’t allow the style kit styles to be optional, which in all(?) other cases makes sense. Since it’s therefore unlikely the theme will support disabling these default styles, the include is overriden in this site to omit the base styles.